实现解释器

+++ jvmgo/ch05/instructions/math/iinc.go
@@ -6,13 +6,13 @@ import "jvmgo/ch05/rtda"
 // Increment local variable by constant
 type IINC struct {
 	Index uint
-	Const int8
+	Const int32
 }
 
 // 从字节码中提取操作数
 func (self *IINC) FetchOperands(reader *base.BytecodeReader) {
 	self.Index = uint(reader.ReadUint8())
-	self.Const = int8(reader.ReadInt8())
+	self.Const = int32(reader.ReadInt8())
 }
 
 // 执行指令, 从局部变量表中读取变量，加上常量值，再写回
